go: use go 1.22 as the default version
In the 1.22.0 release Bottlerocket updated the version of Go used to
build most Go components to Go 1.23, including all versions of
kubernetes prior to k8s 1.30. Go 1.23 had an issue where
a failed call to os/exec would leak a file descriptor. A tight retry
loop in kubelet that called exec and failed very quickly lead to leaked
file descriptors.

https://github.com/bottlerocket-os/bottlerocket/issues/4201

In an attempt to be more conservative with our Go updates, use the lower
of the two minor versions that Bottlerocket supports as the default in
the SDK. This change will coincide with updates to the Go packages in
the bottlerocket-core-kit to use the default Go 1.22 provided by the
SDK, and only override to the latest Go version if necessary.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>