go: build AWS-LC for glibc and musl
Set up Go roots that differ based on the C library that the AWS-LC
syso was linked against.

Fix the "set_cross_go_flags" and "set_cross_go_flags_static" macros
to set GOLIBC. One of these macros is expected to be invoked prior to
any Go build.

Update the Go wrapper scripts to check the value of "GOLIBC" in the
environment and pick the corresponding Go root.

Signed-off-by: Ben Cressey <bcressey@amazon.com>